requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.6
backports.tarfile==1.2.0
beartype==0.22.9
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
cryptography==46.0.3
cyclopts==4.5.0
dataclasses-json==0.6.7
diskcache==5.6.3
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fakeredis==2.33.0
fastapi==0.128.0
fastmcp==2.14.4
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-huggingface==1.2.0
langchain-ollama==1.0.1
langchain-text-splitters==1.1.0
langsmith==0.6.4
lupa==2.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.25.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.1
ollama==0.6.1
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
packaging==25.0
pathable==0.4.4
pathvalidate==3.3.1
pip==24.0
platformdirs==4.5.1
prometheus_client==0.24.1
propcache==0.4.1
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0
pydocket==0.16.6
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.6.0
pyperclip==1.11.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
redis==7.1.0
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==65.5.0
shellingham==1.5.4
sortedcontainers==2.4.0
SQLAlchemy==2.0.46
sse-starlette==3.2.0
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==4.57.6
typer==0.21.1
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
websockets==16.0
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0